+++ packages/polaris-viz/src/components/VerticalBarChart/tests/Chart.test.tsx
@@ -65,6 +65,7 @@ const MOCK_PROPS: Props = {
     fixedWidth: false,
     labelFormatter: (value) => `${value}`,
     integersOnly: false,
+    maxYOverride: 1,
   },
   type: 'default',
   showLegend: false,
@@ -85,6 +86,17 @@ describe('Chart />', () => {
     const multiSeriesBarChart = mount(<Chart {...MOCK_PROPS} />);
 
     expect(multiSeriesBarChart).toContainReactComponent('svg');
+  });
+
+  it('passes maxYOverride as a prop', () => {
+    const chart = mount(
+      <Chart
+        {...MOCK_PROPS}
+        yAxisOptions={{...MOCK_PROPS.yAxisOptions, maxYOverride: 1}}
+      />,
+    );
+
+    expect(chart.prop('yAxisOptions').maxYOverride).toStrictEqual(1);
   });
 
   describe('XAxis', () => {